Modules

Datasets

Datasets.alldatasets — Method

Helper function to return a list of existing datasets

Datasets.downloader — Method

This function takes a url and a destination and downloads the file there with a name specified.

url: The url to be downloaded
dest: Path
fname: File name
```julia
downloader("http://www.julialang.org/", "/tmp/","index.html")
```

Datasets.get_data — Method

Main function to download a standard dataset. Takes input as name from the above list, the destination and a boolean to check if the user wants to extract the data. The filename is inferred from the destination and the url.

get_data("mnist", "/tmp/", 1)

If not found, the list of available datasets is returned.

Datasets.getext — Method

Helper function to get the extension of the datasets. Takes into account patterns like .tar, .tar.gz , .zip etc..

Classification

classification.examplelabeller — Method

A function to show a custom labeller function as an example

This splits the path based on "/" and returns the lower case of the last element

classification.get_img_paths — Function

Get all images in folder

classification.labelFromPattern — Method

Takes paths generated from load_classes as input

Takes a function defining a custom label

eg: labeller(x) = split(x, "/")[-1]

Returns labels based on function specified

classification.load_classes — Method

Identify the number of classes in the data. Return an array with all image paths and class labels

Does not actually load the images. Just returns paths

Assuming heirarchy of classes to be:

Main Folder

class1 -img1 -img2...

class2 -img1 -img2...

classification.load_train_imgs — Function

Load all images

Resize them to given size

one hot encode labels

classification.loadim — Function

Input path -> Read an image -> Convert to array -> Resize to given dimensions -> Convert to float

Available Datasets

Dataset	Link

mnist	Source

cifar10	Source

cifar100	Source

birds	Source

caltech101	Source

pets	Source

flowers	Source

food	Source

cars	Source

imagenette	Source

imagenette320	Source

imagenette160	Source

imagewoof	Source

imagewoof320	Source

imagewoof160	Source

imdb	Source

wikitext103	Source

wikitext2	Source

wmt	Source

ag	Source

amazon	Source

p-amazon	Source

dbpedia	Source

sogou	Source

yahoo	Source

yelp	Source

p-yelp	Source

camvid	Source

pascal	Source

hmbd	Source

ucf	Source

kinetics700	Source

kinetics600	Source

kinetics400	Source